Further dev work for #3, #8 and #9
The operations of the batcher appear unaffected with this implementation, which is great. Dynamic import of the cluster backend is operational but it seems there is work to do to evaluate how the dummy backend works. Once verified I'll either close the issues or provide a commit concluding the particular issues, whichever is best.