Fixes #31: detection error propogating back to batcher
This was actually fine behaviour for what was there originally, but find_id for slurm specific had a catch all that didn't cover the string parsing. Therefore, this has been refined to avoid the bad practice of catching everything and the ValueError caught locally which now covers the slurm specific behaviour.