Finished #26 - added auto skip on failure to submit
To try and alleviate the potential for repeat resulting in endless looping I've added
an auto-skip functionality to the run_batch_item logic. This is good as it's the precursor,
now that return values from this method are usable, to evaluating the output state of the
job. This has been tested with the dummy runner so far but should be equally portable to
the other cluster implementation